Move SerializedPage to SPI
