+++ presto-main/src/main/java/com/facebook/presto/execution/buffer/BufferResult.java
@@ -13,6 +13,7 @@
  */
 package com.facebook.presto.execution.buffer;
 
+import com.facebook.presto.spi.page.SerializedPage;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
