+++ presto-main/src/main/java/com/facebook/presto/server/AsyncPageTransportServlet.java
@@ -19,8 +19,8 @@ import com.facebook.presto.execution.TaskId;
 import com.facebook.presto.execution.TaskManager;
 import com.facebook.presto.execution.buffer.BufferResult;
 import com.facebook.presto.execution.buffer.OutputBuffers.OutputBufferId;
-import com.facebook.presto.execution.buffer.SerializedPage;
 import com.facebook.presto.operator.ExchangeClientConfig;
+import com.facebook.presto.spi.page.SerializedPage;
 import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.ListenableFuture;
 import io.airlift.units.DataSize;
