+++ presto-main/src/main/java/com/facebook/presto/server/PagesResponseWriter.java
@@ -13,8 +13,8 @@
  */
 package com.facebook.presto.server;
 
-import com.facebook.presto.execution.buffer.SerializedPage;
 import com.facebook.presto.spi.Page;
+import com.facebook.presto.spi.page.SerializedPage;
 import com.google.common.reflect.TypeToken;
 import io.airlift.slice.OutputStreamSliceOutput;
 import io.airlift.slice.SliceOutput;
