+++ presto-main/src/test/java/com/facebook/presto/execution/buffer/TestThriftBufferResult.java
@@ -14,6 +14,7 @@
 package com.facebook.presto.execution.buffer;
 
 import com.facebook.presto.spi.page.PageCodecMarker;
+import com.facebook.presto.spi.page.SerializedPage;
 import com.google.common.collect.ImmutableList;
 import org.testng.annotations.Test;
