+++ presto-main/src/test/java/com/facebook/presto/operator/MockExchangeRequestProcessor.java
@@ -22,8 +22,8 @@ import com.facebook.presto.client.PrestoHeaders;
 import com.facebook.presto.execution.buffer.BufferResult;
 import com.facebook.presto.execution.buffer.PagesSerde;
 import com.facebook.presto.execution.buffer.PagesSerdeUtil;
-import com.facebook.presto.execution.buffer.SerializedPage;
 import com.facebook.presto.spi.Page;
+import com.facebook.presto.spi.page.SerializedPage;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
