+++ presto-spi/src/main/java/com/facebook/presto/spi/page/SerializedPage.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.execution.buffer;
+package com.facebook.presto.spi.page;
 
 import io.airlift.slice.Slice;
 import org.openjdk.jol.info.ClassLayout;
